2nd commit

+++ database/migrations/2019_11_05_103537_create_orders_table.php
@@ -16,7 +16,7 @@ class CreateOrdersTable extends Migration
         Schema::create('orders', function (Blueprint $table) {
             $table->increments('id');
             $table->string('name');
-            $table->integer('price');
+            $table->decimal('price', 5, 2);
             $table->integer('quantity');
             $table->timestamps();
         });
